+++ tests/testthat/man/testdir/DEV/functions/inc1.R
@@ -1,0 +1,3 @@
+#' @include inc2.R inc3.R
+inc1 <- function(){
+}