+++ tests/testthat/man/testdir/QA/functions/inc2.R
@@ -1,0 +1,3 @@
+inc2 <- function(){
+  print("inc2")
+}